Adds a Map Changelog Entry Category (#87034)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This should hopefully cut down on mappers using the add:/fix:/del: tag
I've talked to the maptainers about this and they seem to not be
concerned about gbp in their own system, so we should be trying to push
mappers away from using add/remove/(maybe? closing issues is cool) fix
tags and such to avoid overlap.

I've intentionally not linked this to a label due to the Map Edit label
autodetecting anyway. Should be directed only at players.

This does mean less label resolution on mapping prs, not sure how I feel
about that but I didn't want to duplicate the old image
add/remove/change labels.

Requesting wider input on this and also pinging to notify.
@tgstation/commit-access @tgstation/maptainers 

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
map: added a map specific changelog entry
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->