[No GBP] Minor change/fix to MI13 Agent's loadout (#87027)
## About The Pull Request
The agents had a 10% chance to spawn with a machine trapper. Now they
have a 10% chance to spawn with thermal goggles (non-chameleon).
## Why It's Good For The Game
Turns out the machine trapper doesn't do anything when a non-traitor
uses it. The replacement item is thermals, it's on-theme for the faction
and is about the same power as the other items in the pool.
## Changelog
:cl:
balance: MI13 Fugitive Hunters can no longer spawn with machine
trappers, instead they may now spawn with thermal goggles.
/:cl: